# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/",
    "paths": {
        "/v1/transactions": {
            "post": {
                "description": "Get transactions by filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Количество записей на одной странице",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Страница списка",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "description": "Фильтр для транзакций",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.filterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.WrapperResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/v1.transactionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    }
                }
            }
        },
        "/v1/transactions/top-up": {
            "post": {
                "description": "Top up balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Top up balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.WrapperResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/v1.transactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    }
                }
            }
        },
        "/v1/transactions/transfer": {
            "post": {
                "description": "Transfer money between users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Transfer money",
                "parameters": [
                    {
                        "description": "Данные для перевода средств",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.remittanceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.WrapperResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "v1.WrapperResponse": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "description": "Код запроса",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "Запрашиваемые данные при успешном запросе"
                },
                "message": {
                    "description": "Сообщение об ошибке",
                    "type": "string",
                    "example": ",Все пропало"
                }
            }
        },
        "v1.filterRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "v1.remittanceRequest": {
            "type": "object",
            "required": [
                "recipient_id",
                "sender_id",
                "sum"
            ],
            "properties": {
                "recipient_id": {
                    "type": "integer",
                    "example": 2
                },
                "sender_id": {
                    "type": "integer",
                    "example": 1
                },
                "sum": {
                    "type": "string",
                    "example": "100"
                }
            }
        },
        "v1.transactionResponse": {
            "type": "object",
            "required": [
                "comment",
                "date",
                "id",
                "sum",
                "type",
                "user_id"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "Пополнение баланса"
                },
                "date": {
                    "type": "integer",
                    "example": 21212
                },
                "id": {
                    "type": "integer",
                    "example": 12
                },
                "sum": {
                    "type": "string",
                    "example": "100"
                },
                "type": {
                    "type": "string",
                    "example": "RECHARGE"
                },
                "user_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}